Some World 2 and World 3 Stages

+++ CatMobileOfficial3.gmx/scripts/scr_reset.gml
@@ -1,2 +1,27 @@
-//resets the game
-file_delete("Save.sav");
+if (file_exists("Save.sav")) file_delete("Save.sav");
+global.w1s1 = 0;
+global.w1s2 = 0;
+global.w1s3 = 0;
+global.w1s4 = 0;
+global.w1s5 = 0;
+global.w1s6 = 0;
+global.w1s7 = 0;
+global.w1s8 = 0;
+
+global.w2s1 = 0;
+global.w2s2 = 0;
+global.w2s3 = 0;
+global.w2s4 = 0;
+global.w2s5 = 0;
+global.w2s6 = 0;
+global.w2s7 = 0;
+global.w2s8 = 0;
+
+global.w3s1 = 0;
+global.w3s2 = 0;
+global.w3s3 = 0;
+global.w3s4 = 0;
+global.w3s5 = 0;
+global.w3s6 = 0;
+global.w3s7 = 0;
+global.w3s8 = 0;
